ilb: force cilium service reconciliation when node labels change
Currently, changing the labels of a K8s Node, doesn't trigger
a Cilium Service reconciliation. This only happens when the Agent
gets restartged or the K8s Service gets modified. At least this is
the case for the "old" non-experimental LB mode.

As a temporary workaround (until we use the new experimental LB mode),
this commit changes the ILB reconciliation to enforce a reconciliation
of the Service in Cilium.

This is achieved by adding a new K8s Annotation on the K8s Service
that gets changed whenever the T1 nodes change (hashed T1 IPs).

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>