rib: Introduce a simple RIB
Introduce the Routing Information Base which BGP Control Plane (or any
other Cilium subsystems) can write routes to it. This commit only covers
very basic data structure. The following is an overview of how it works.

The RIB data structure consists of per-VRF trie for supporting L3 VPN
routes. The key of the trie is a prefix and the value of the trie is a
structure called Destination.

The Destination represents a destination network and possible routes to
reach to there. The routes are differentiated with the owner. This
structure allows route owners (e.g. different BGP instances) to
reconcile on the set of desired routes individually even they are
competing with the same prefix.

The RIB takes care of selecting the best route out of multiple routes in
the same Destination. The best path selection algorithm will be
implemented in the subsequent commit.

When there's a change in the best path (including the first insertion
and the last deletion), the DataPlane will be notified. The subsequent
commit will add a DataPlane and notification support.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>