bpf: host: respect ENABLE_IDENTITY_MARK when looking at MARK_MAGIC_IDENTITY
[ upstream commit 2566ce2b451e4a74b75818e59046175219bc07b3 ]

For environments where option.EnableIdentityMark is not set (and thus
from-container doesn't set the source identity in the skb->mark), we
shouldn't attempt to retrieve the identity.

Which leaves the question whether all the *other* use of skb->mark in this
program is safe. Or all those features are simply incompatible with
running in a CNI-chained environment, where skb->mark is not owned by
Cilium (see https://github.com/cilium/cilium/pull/12185).

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>