ci: Populate image versions in CLife build and conformance tests
Make use of the script that populates the version of the images
of the operands: Cilium agent, operator, envoy etc in values.yaml
when the CLife image is built in CI and the conformance tests are
runs.
This also makes redundant the population of the versions in the
custom resource.

Signed-off-by: Frederic Giloux <frederic.giloux@isovalent.com>