hubble-ee: fix exporter bytes total metric fail to init
From WrapWriter, when the metrics handler was not initialized
we would avoid wraping the exporter writer and return the provided
writer as-is. This means that if the handler after the call to
WrapWriter, we could not recover and fail to update the bytes total
metric.

To fix, provide a handle for the metrics handler to the bytecounter
writer instead of the prom metric handle. Move the check for
whether the handler is initialized in a method on the handler, and
call that method from the bytecounter writer instead, allowing us
to make the decision for each write instead of at init time.

Signed-off-by: Alexandre Barone <abalexandrebarone@gmail.com>