images: build UBI startup script
It was wrongly assumed that CILIUM_NODEINIT_VERSION content
in install/cilium/Makefile.values was a git sha of cilium/image-tools.
It was actually a docker tag.

This change changes enterprise startup script to treat it as a docker
tag of the image from which it needs to extract manage-startup-script
file from.

Signed-off-by: Maciej Kwiek <mkwiek@cisco.com>