srv6: Introduce SRv6 RIB DataPlane implementation
Introduce a new rib.DataPlane implementation backed by an existing SRv6
eBPF datapath. It consumes the RIB update event and writes to the
PolicyMap or SIDMap depending on the NextHop type. If the NextHop is
H.Encaps, it writes to the PolicyMap. If the NextHop is End.DT4, it
writes to the SIDMap. This hides this map separation (they could be
consolidated into the single FIB-ish structure) from the RIB layer.

The unit test covers the basic Create/Update/Delete scenario as well as
some error cases. It uses mocked BPF maps for unprivileged test for
quick local testing and uses real BPF maps for privileged test. In this
way, we can cover both cases with CI.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>