bgpv2,rr: Don't change nexthop for non-RR eBGP peer
Set nexthop-unchanged action for non-RR eBGP peer (not an RR or RR
Client) of the route reflector instance. This allows implementing a
strawman's third party nexthop.

Note that this is not how third party nexthop works in general. In the
real third party nexthop implementaion, we need to figure out whether
the nexthop is l2 reachable or not and change/unchange nexthop depends
on it. However, the current implementation should work for most of our
usecases.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>