egressgwha: Exercise update from default route in test
Change the starting Destinations and the updated ones in TestOps to
exercise a removal of the "stale" IPv4 default route in the Update
reconcilation op.

This highlights an issue with an incorrect conversion from net.IPNet to
net.Prefix that leads to a netlink error when deleting the route:

```
=== RUN   TestOps
    ops_linux_test.go:146:
        	Error Trace:	/home/pippolo/isovalent/cilium/enterprise/pkg/egressgatewayha/egressipconf/ops_linux_test.go:146
        	Error:      	Received unexpected error:
        	            	failed to delete route for egress IP 192.168.1.50 and interface dummy0: source and destination ip are not the same IP family
        	Test:       	TestOps
        	Messages:   	expected no error from second update
```

The subsequent commit fixes the issue.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>