egressgwha: Fix conversion from net.IPNet to netip.Prefix
The egw IPAM reconciler takes care of reconciling the egress addresses,
rules and routes in accordance with what is specified in the IEGP
policies.  To do so, the Update and Delete operations list all the
installed routes with safenetlink.RouteListFiltered.

When a route has been installed for a destinationCIDR equal to the
default v4 route, that is, "0.0.0.0/0", RouteListFiltered returns a
net.IPNet that contains an IPv4 equal to the v4 default route but in
16-byte form. The subsequent conversion to a netip.Prefix performed by
ipNetToPrefix returns an IPv6 prefix that ultimately lead to the
following netlink error when the route has to be deleted:

`failed to delete route for egress IP 10.2.0.1 and interface net0: source and destination ip are not the same IP family`

The commit fixes the conversion using netipx.FromStdIPNet.

Related: https://github.com/isovalent/customer-support/issues/1395

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>