ilb: introduce k8sServiceRef on LBBackendPool CRD
This commit adds the possibility to reference K8s Services in
the CRD `LBBackendPool`.

Therefore, a new backend type `K8sService` has been introduced. When
this type is configured, the respective backends must define
the endpoints by referencing a K8s Service by its name (currently
in the same K8s namespace only).

```
apiVersion: isovalent.com/v1alpha1
kind: LBBackendPool
metadata:
  namespace: default
  name: lb-1
spec:
  backendType: K8sService
  backends:
    - k8sServiceRef:
        name: test1
      port: 8080
    - k8sServiceRef:
        name: test2
      port: 8080
```

Note: The backend port refers to the port of the K8s Service (`port`), and not the
`targetPort` (that would be the one of the `Pod`).

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>