ilb: validate K8s Service reference and port on LBBackendPool status
This commit adds validation logic to the `LBBackendPool` reconciler.
It validates that all referenced K8s Services do exist and that the
referenced port is defined in `spec.ports`.

If not valid, this will populate the status of the `LBBackendPool` resource
which will block the reconciliation of all `LBService`'s that are referencing
the invalid `LBBackendPool`.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>