rib,srv6: Provide RIB and DataPlane cells
Provide RIB and DataPlane cells to the enterprise daemon. SRv6 DataPlane
constructor is adjusted to provide only when SRv6 is enabled.

Since the RIB is an independent module, we have a configuration
combination that DataPlane is not provided (nil), but RIB is provided.
To preventing panicing, provide no-op DataPlane implementation when nil
DataPlane is provided.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>