rib: Introduce rib/list Hive Script command
Introduce a new rib/list Hive Script command that dumps RIB entries. It
will be provided with the function ribReadCommands. We named it as read
because at runtime, we only let users to read the RIB through Hive
Script. The write variant will be implemented in the future for the
testing purpose.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>