ilb: translate backend port via service ports (port -> target port)
If backend typep `K8sService` is used, the backend port points to the
port of the Service (`port` field) and not the `targetPort` (or Pods port).

Therefore, this commit translates the service port to the Pod port by
looking up the target port in the `Service` resource. If the target port
is defined by its name - the logic also needs to resolve the actual
port of the Pod by looking it up in the corresponding `EndpointSlice`'s.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>