service: Revert variable declaration format change
This line was changed in main-ce but not in the upstream version of the
commit, which caused the two to diverge. The formatting doesn't matter
either way, so this commit just reverts the internal change.

No functional difference.

Related: b8725f5b3b0a ("service: notify health observers about deleted service")
Signed-off-by: Joe Stringer <joe@isovalent.com>