.github: Revert main-ce distro check
Commit 4d286ddfbae8 ("integration: Update condition based on runner name")
internally modified this check to solve the same problem as the upstream
commit 95018178c1da ("integration: Update condition with startsWith")
aimed to solve, which caused this workflow to diverge and have
conflicts. Since upstream has a fix for this, just use the upstream fix
and revert the internal change.

I checked the value of `GH_RUNNER_EXTRA_POWER_UBUNTU_LATEST` and it
should continue to match this statement regardless of this change, so
there should be no functional difference to testing with this commit.

Signed-off-by: Joe Stringer <joe@isovalent.com>