bgpv2,srv6: Introduce a new importVPNRouteReconciler
Introduce a new importVPNRouteReconciler which replaces an existing
srv6-responder-based route import implementation. Instead of relying on
the "responder", now it can directly import the routes from BGP RIB to
Cilium's internal RIB.

It still leverages the existing status reconciler infrastructure. This
works, but has a few caveats:

1. It's very inefficient. Since the status reconciler doesn't support an
   incremental update, we do full reconciliation for every route update.
   This should work well if the number of imported routes from the peer
   is small, but for large number of routes, it may have a problem.

2. It cannot support ECMP because of the GoBGP's spec. We use GetRoutes
   API for this implementation. GoBGP doesn't provide the multi path
   information with this API. We instead need to use Watch API to
   support that. It's ok for now since our datapath cannot handle ECMP
   anyways, but we should support that in the future.

Because of the above caveats, we should migrate to the Watch API based
implementation in the future.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>