fqdn-ha: proxy: Delay opening socket until rules are received
We need to get the set of allowed FQDN names per-endpoint from the
agent. Until we receive these, all DNS requests will be spuriously
denied, as we don't yet have that endpoint's configuration.

Unfortunately, as soon as we open the DNS socket, we will start
receiving approximately 50% of DNS requests. So, delay opening the
socket until we have received at least 1 endpoint.

This means that for clusters with no DNS interception configured, we
will never open the socket.

Signed-off-by: Casey Callendrello <cdc@isovalent.com>