ilb: add e2e test for label based backend selection with k8sServiceRef
This commit adds a simple E2E ILB test for the label-based backend selection
functionality where a LBBackendPool references its backends via a reference
to a K8s Service.

The test checks that requests succeed and are loadbalanced to all pods of the
referenced K8s Service. It's tested for  T1-only & T1&T2 with a TCPProxy application.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>