ci: Disable enterprise-clustermesh-overlapping-podcidr
Based on the discussion in the following thread, we broke the
Overlapping PodCIDR with upstream change.

https://isovalent.slack.com/archives/C03AHFW7X53/p1745291397808219

Since the feature is beta and missing the user, the commits caused the
breakage were merged as is with
https://github.com/isovalent/cilium/pull/7067.

The CI was removed marked as non-required here:

https://isovalent.slack.com/archives/C02HCGENRF0/p1745494602368679

However, since it fails 100% in the main-ce and shows up for every PR's
CI run, it is better to disable it until we fix it for not waisting GH
resource fee and prevent developers' confusion.

The issue is tracked here
(https://github.com/isovalent/cilium/issues/7178).

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>