rib: Introduce Kubernetes protocol
Introduce a new protocol Kubernetes which typically used for the routes
originated from CRD. We may want to adjust the AD values in the future,
but for the moment, we assign 10. The value is chosen based on the range
which is lower than BGP, but still have some room for inserting higher
priority protocol.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>