contrib/mindful: Add mindfulness check script
This script compares the current HEAD against the most recent merge of
the upstream OSS Cilium branch in order to highlight any changes that
are made to files in the OSS tree. The intent is to minimize the number
of direct git conflicts caused by merging such code changes into the
enterprise. With the results of this script, it will be easier for
developers and reviewers to identify which changes should be upstreamed
or which we may want to consider as exceptions to the rule (and carry
the burden of diverging from upstream).

Signed-off-by: Joe Stringer <joe@isovalent.com>