ilb tests: add a test filter with negation
Add a flag to filter the tests that are run during the Isovalent
LoadBalancer test suite.

Providing the flag multiple times will allow multiple expressions to
match. A test which matches any of the expressions will be run.

Expressions that start with '!' will not be executed. If a test name
matches any of these negated expressions, it will not run, regardless of
the number of non-negated expressions that it matches.

Signed-off-by: Andrew Sauber <andrew.sauber@isovalent.com>