(cee) bpf: cilium-mesh: remove SKIP_POLICY_MAP
This macro doesn't exist anymore in-tree, so we don't have to deal with
it.

Glancing at the code it also never made sense to use it here in the
first place. Contrary to what the naming suggests, the macro was *not*
about network policy enforcement. But about whether a program can
utilize the `cilium_call_policy` map, for tail-call style delivery to a
local endpoint.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>