ilb: don't add TLSinspector listener filter for TCPProxy app
Currently, the `TLSInspector` listener filter is added to each
TCP listener in Envoy.

The reason for doing this is to differentiate between different
application types that could potentially be configured on the same
`LBService` - e.g. HTTPProxy and HTTPSProxy.

But this can also lead to some issues with some protocols - e.g.
mysql. See https://github.com/envoyproxy/envoy/issues/21044 for
more information.

Therefore, this commit removes the TLSInspector listener filter
if TCPProxy application is configured. In these cases there's never
(even theoretically) any other application type present on the same
listener.

The only reason might be if we introduce T1->T2 health checks that
use HTTPS instead of plain HTTP. But we might be able to only rely on
the source IP (which would be a T1 node IP) in these cases (we already
use this filter).

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>