ci: replace schedule trigger with merge_group
The commit improves the external images trigger workflow by replacing the
`schedule` trigger with the `merge_group` trigger.

This is required because of the merge queue approach.

When a pull request is added to the merge queue, the changes in the pull
request are grouped into a merge_group with the latest version of the
base_branch as well as changes from pull requests ahead of it in the queue.
GitHub will merge all these changes into the base_branch once the checks
required by the branch protections of base_branch pass.

Signed-off-by: viktor-kurchenko <viktor.kurchenko@isovalent.com>