fix: Have the related images populated in CLife ClusterServiceVersion
opm tooling does not transfer the relatedImages field from
the base template to the rendered ClusterServiceVersion.
They are now populated as environment variables on the
deployment for getting the expected result.
The CLife image needs to be added to the related images as well.
It is only possible with digest pinning after it has been built.
That's the reason for splitting the set-images.sh script into two parts
that can be run at different point in the build process.

Signed-off-by: Frederic Giloux <frederic.giloux@isovalent.com>