egressgwha: simplify error handling when selecting egressIP
Instead of setting `EgressIPNotFoundIPv*` from every error path, just set
it by default. And later override it with the actual egressIP, once we
were able to derive the gateway node's network configuration.

This helps to ensure that if we were not able to derive a network config
on the gateway node, the policy entry always ends up with the
EgressIPNotFoundIPv* value - and thus explicitly drops any EGW traffic that
is forwarded to the gateway.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>