Move allowedCommands into self-hosted renovate configuration
At some point an enterprise-renovate-config.js was introduced, which
separates the self-hosted renovate configuration from the repository
configuration. `allowedCommands` is part of the self-hosted
configuration but wasn't moved to the new config file when it was added,
meaning the allowedCommands defined in the renovate.json5 has no effect.

To fix it means to move any self-hosted renovate config options into the
new enterprise-renovate-config.js.

Signed-off-by: Chance Zibolski <chance.zibolski@gmail.com>