rib,command: Determine the best path based on the route equality
Currently, the rib/list command determines the best path based on the
pointer equality. This is error prone, so determine the best path based
on the equality between Destination.best.

Currently we have a bug that this misbehaves when we upsert exactly the
same route twice. Due to the change in the #7299, we no longer update
the Destination.best if the new route and the existing best path is
equal. However, we do update the existing entry in the
Destination.routes, so the pointer changes. Therefore after the second
upsert, the rib/list command starts to miss the best path.
Equality-based approach is slightly inefficient than the pointer-based
approach, but avoids this kind of the bug.

Fixes: #7299

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>