fix: Add hostnetwork-v2 SCC privilege to CLife
The namespace where cilium is installed on OpenShift has an
annotation giving its service accounts elevated privileges.
In Red Hat certification pipeline this namespaces does not
however get created through our manifests and this annotation
is missing, inducing the failure of the certification pipeline.
This change adds access to the hostnetwork SCC to CLife, also
when the installation namespace does not have the annotation.

Signed-off-by: Frederic Giloux <frederic.giloux@isovalent.com>