rib: Add initial garbage collection mechanism
We have a scenario where some routes are withdrawn or some route owners
are gone during the agent restart. That leaves some stale routes to the
DataPlane. We need a GC mechanism to cleanup such routes.

In this commit, we introduce a GC mechanism that works like following.

1. At the agent startup, RIB imports the currently available routes from
   DataPlane. All routes will be imported with unknown protocol which
   always loses the best path selection.

2. Route owners will be appeared eventually and fill up the RIB with
   their routes.

3. After 5min (by default), the garbage collection mechanism deletes all
   owner=unknown routes assuming all route owners are fully synced at
   that point. This still leaves the active route in the DataPlane, but
   deletes the stale routes.

We could also wait for the actual sync of all route owners. However,
that's gonna be a bit tricky because the route owners are dynamic (e.g.
BGP CPlane Instances), so we need a synchronization mechanism for
waiting for the initial owner list. And then, we also need an another
synchronization mechanism for initial route list for each owners.

To avoid introducing such complexity, we instead rely on this simple
timeout-based approach. It has a certain risk that we lose the
connectivity because we run GC too early or having the undesired
connectivity because we don't run GC quickly. For those who want to
adjust the timeout, add a hidden configuration knob as an escape hatch.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>