bgpv2: Set LocalASN field and protocol for the expected routes properly
In the #7342, we have introduced a proper iBGP vs eBGP determination in
the importVPNRouteReconciler. As a result, it starts to rely on the
DesiredConfig.LocalASN field and also started to set a proper iBGP or
eBGP protocol to the routes imported into the RIB.

Due to the merge race, #7326 introduced a test breakage because the
dummy reconciliation parameter in the replay test doesn't set LocalASN
field and the protocol of the expected routes are still "unknown". Fix
it.

Fixes: #7326

Suggested-by: Rastislav Szabo <rastislav.szabo@isovalent.com>
Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>