hubble-k8s-exporter: Disable cilium identity export
The hubble k8s exporter started exporting cilium identities by default.
This is currently failing for integrated timescape because it doesn't
have sufficient permissions to do so. This doesn't impact functionality,
but spams error logs in the exporter.

The cilium identities are currently not used by integrated timescape, so
this change disables cilium identities export to avoid the log spam. We
can re-enable it again and add the necessary permissions, once we
actually need them.

Signed-off-by: Fabian Fischer <fabian.fischer@isovalent.com>