bgpv2: Use frontends table for health information
Use the Table[*loadbalancer.Frontend] for the backend health info instead
of the deprecated ServiceHealthCheckManager.

Refactor the script tests to use the load-balancing control-plane for
setting health and convert to EndpointSlices.

We will eventually want to remove the uses for the service and endpoints
"DiffStores" and just use the frontends table. This is a bigger refactoring
for later (#7380).

Signed-off-by: Jussi Maki <jussi@isovalent.com>