timescape: Add experimental clustermesh support
Currently, integrated Timescape doesn't support clustermesh at all,
while Hubble Relay supports clustermesh by requesting flows from all
local and nodes remote nodes.

This change implements experimental clustermesh support for integrated
Timescape similarly, by deploying Timescape in a single "primary"
cluster and only deploying Vector and the hubble-k8s-exporter in all
other clusters and point a global service to the primary Timescape
instance.

This has some obvious scaling issues and creates a lot of cross cluster
traffic, so it's not suitable for every clustermesh deployment. It can,
however, be a good solution for deployments with one main cluster and a
few small satellite clusters. For example when connecting to TGWs, and we
pan to use it for the Isovalent Network Bridge.

I'm marking it as experimental for now, and it's not yet intended to be
used by customers.

Signed-off-by: Fabian Fischer <fabian.fischer@isovalent.com>