bgpv2: Remove unnecessary logs in SRv6LocatorPoolReconciler
When there's a PeerConfig with IPv4 unicast address family configured
and the Advertisement for locator pools are selected with that, the
SRv6LocatorPoolReconciler produces this kind of the warnings for every
reconciliation.

```
time="2025-06-07T02:45:26Z" level=warning msg="Invalid address family for this advertisement type, skipping" advertisement_type=SRv6LocatorPool family=ipv4 subsys=enterprise-bgp-control-plane
```

Actually, this is possible depending on the user's configuration. For
example, when they are using the same labels for SRv6LocatorPool
Advertisement and the rest of the IPv4-available advertisements.

To suppress this warning, users need to separate the label for
SRv6LocatorPool and others which is not a good UX. Since this case is
expected and harmless, let's just remove this log.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>