hubble: Set same default config in dynamic and static exporter
When configuring the dynamic exporter, the parsed configuration
is different that when using static config This is because when using
non-pointer types in structs to be unmarshaled, we can't differentiate
between unset or set to the default value for the type.

We currently use a mix of pointer and non-pointer types in the struct
used for the yaml-based config file. To make parsing easier, introduce
a new intermediate type with all fields using pointers and update the
existing FlowLogConfig type to use non-pointers for all fields. The new
intermediate type has a method that can convert to a slice of
FlowLogConfigs and set the default values based on `defaultConfig`,
the same defaults the static exporter is based on.

Signed-off-by: Alexandre Barone <abalexandrebarone@gmail.com>