chore: Build CI versions of the UBI based external images
Some images, like envoy and certgen are built outside of the
enterprise repository. Their release versions of the UBI based
flavor were built in CI when the reference for the helm charts
was changed. This first commit does so that a CI version instead
of a release version gets built and pushed to isovalent-dev.
A subsequent commit will look at generating the release versions.
This will avoid:
- possible race conditions where more than one release version
gets built for the same tag
- having the need for the release manager to approve CI runs

Signed-off-by: Frederic Giloux <frederic.giloux@isovalent.com>