chore: Use release tag for external image in CLife
The OSS tag was used so far for references to external images,
e.g. envoy, certgen in CLife manifests, i.e the internal helm
values and the ClusterServiceVersion. Now the release tag,
e.g. v1.16.10-cee.1 is used instead. This aligns the tags for
the references to UBI images in CLife manifests to a change
introduced in the tags for the references in the helm charts,
which happens after the CLife image has been built.

Signed-off-by: Frederic Giloux <frederic.giloux@isovalent.com>