chore: Build release versions of the UBI based external images
Some images, like envoy and certgen are built outside of the
enterprise repository. Their release versions of the UBI based
flavour were built in CI when the reference for the helm charts
was changed. This moves the build to release time. A previous
commit has added the build of CI images.

Signed-off-by: Frederic Giloux <frederic.giloux@isovalent.com>