ilb: ipfix: implement in-bpf aggregation, add ipv6 and l2 flows
Make the backend for flow logs more scalable and also implement IPv6
and non-IP reports.

For now non-IP only include src and dst mac, and ethernet payload type.
It also is not exported using IPFIX, only stdout, as IPFIX doesn't
standartize L2.

To make things scalable the following approach is imlemented.
The fl_table_in_use defines which of tables, cilium_fl_v4_1 or
cilium_fl_v4_2, to use to update flow logs (for IPv6 and non-IP table
names are cilium_fl_v6_{1,2} and cilium_fl_l2_{1,2}, correspondingly).
This is a global counter which can be equal to

    0: flow logs are disabled
    1: flow logs are written to cilium_fl_v4_1
    2: flow logs are written to cilium_fl_v4_2

Only cilium agent writes to this variable. When a write occurs,
before summarizing the old table, agent waits for a period of time
long enough that all the writes to the old table are finished.

Update vendors to remove ringbuf.

Signed-off-by: Anton Protopopov <aspsk@isovalent.com>