enterprise/bgpv2: Refactor service route policy rendering logic
Use common method for rendering service route policies
for all types of svc advertisements instead of 3 different
methods to avoid code repetition. Rely on exiting path
rendering methods (get<svcType>IPPaths) instead of copying
the same logic in the policy rendering methods.

Also fix prefix rendering in get<svcType>IPPaths to properly
zero masked-out bits, e.g. to produce 192.168.1.0/24 instead
of 192.168.1.1/24. This was not causing any issues (zeroed on
the gobgp level), but as we now reuse it for policy rendering,
it needs to be fixed.

Signed-off-by: Rastislav Szabo <rastislav.szabo@isovalent.com>