enterprise: initial scaffolding of the enterprise cmapisrv script tests
Create an initial version of the enterprise clustermesh-apiserver script
tests, similar to the OSS one. Currently, we only run the OSS tests
themselves, as a sanity check to validate that we are not breaking
anything in the enterprise version. We need to skip the tests asserting
the correctness of the cluster config though, because we intend to use
an extended version with enterprise-specific capabilities, and they
would fail.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>