egressgwha: fallback to fib_lookup based selection for dummy interfaces
The PR #4362 introduced support for ifindex-based BPF forwarding.

However, this broke use cases where Egress IPs are assigned to dummy
interfaces, because the datapath now attempts to redirect_neigh to the
dummy interface, which is not supported.

To fix this, skip populating the ifindex in the policy map when a dummy
interface is specified in the IEGP. This allows the datapath to
fallback to fib_lookup based selection as before.

Signed-off-by: Yusuke Suzuki <yusuke.suzuki@isovalent.com>