ent/makefile: update the CNI binary as part of image-fast-agent target
Similarly to [1], update the cilium-cni binary via the kind-image-enterprise-fast
Makefile target, to simplify iterating on during development. However, the target
path is changed to /opt/cni/bin, as that's where the CNI binary is expected to be
located. I'll follow up with a similar change upstream as well.

It would be nice to simply call the upstream target, to ensure consistency,
but that's currently not possible because an enterprise version is not available
for all binaries (or it is not located in the expected folder). More invasive
adaptations would be required, which are postponed for now.

[1]: a951c39c8ea4 ("make: build and install cilium-cni binary in fast targets")

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>