ilb: fix t1-only health check annotations
Currently, T1-only L4 health checking that performs the actual health checks
to the backends uses the same config values that are used for T1->T2 health
checking. These are wrong, because in that case the config values from the
CRD `LBBackendPool` should be used.

Therefore, this commit is fixing this and writes the correct health check
config values to the T1 k8s service annotations.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>