bgp,rr: Add route-reflector advertisement script test
Add script test case for route-reflector. This test uses a "real" Cilium
BGP Control Plane as a route-reflector and uses GoBGP as clients and
external routers. While we have no real RR clients, this test is useful
for testing if the route-policy rendering and route-reflection are
working properly for route-reflector.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>