ilb: unique sysdump filename per matrix run
Currently, capturing sysdumps on an ILB E2E test failure results in a
sysdump file with the following pattern: `cilium-sysdump-<TESTFUNCNAME>.zip`.

If two runs (matrix) fail on in the same test(-func), the two output files are
the same. This way we miss one of the runs and it potentially results in corrupt
zip files when trying to merge the artifacts.

Therefore, this commit makes the outputfiles unique by passing the name
of the sysdump output filename as parameter and include the matrix name as
part of the filename.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>