enterprise: Initial tooling for agent self-diagnostics
This introduces enterprise/pkg/diagnostics which provides [diagnostics.Registry]
for registering diagnostic conditions. Each condition is a set of metadata
(ID, description, severity) and an evaluation function which has access to
an [diagnostics.Environment] through which metrics can be accessed
(the evaluation functions can access any object if needed):

  func registerMyConditions(reg *diagnostics.Registry) {
    reg.Register(
      diagnostics.Condition{
        ID: "my_condition",
        SubSystem: "my-subsys",
        Severity: diagnostics.SeverityMajor,
        Description: "description of my condition",
        Evaluator: func(env diagnostics.Environment) (msg string, failed bool) {
          m, err := env.Metric("my_gauge", nil)
          if err == nil && m.Value > 1000.0 {
            return "oh no, my_gauge > 1000.0!", true
          }
          return "", false
        },
      },

      // ... more conditions ...
      diagnostics.Condition{ ...},
    )
  }

The registered conditions are evaluated periodically (by default every 5 minutes)
and the combined status is written to a file pointed to by '--diagnostics-export-file'.

Signed-off-by: Jussi Maki <jussi.maki@isovalent.com>