ilb: validate that JWT authentication is configured if JWT req. filtering
This commit introduces additional validation (including status condition)
on the `LBService` CRD. Whenever JWT claim based request filtering is
configured on a HTTP route, JWT authentication needs to be configured on
the same route or globally on the HTTP(S)Proxy application.

This is necessary because it's the Envoy JWT authentication HTTP filter
that extracts the JWT claims once it has been successfully validated.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>