ilb: fix JWT claim based HTTP request filtering
Currently, claim based HTTP requestfiltering doesn't work as expected
if it's configured in combination with source ip based requestfiltering.

The reason is that the resulting envoy RBAC config doesn't configure
an "and-principal" when merging the RBAC principals. This way the individual
RBAC principals are handled in an or-fashion where only one has to match.

Therefore, this commit fixes this by adding an explicit and-principal
(which we already do for the RBAC permissions).

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>