rib,srv6: Fix the deadlock of initial GC
Calling DataPlane.ProcessUpdate while iterating over the datapath with
DataPlane.ForEach is not safe because typically DataPlane.ForEach
implementation holds read lock and the DataPlane.ProcessUpdate holds
write lock. Currently, we're calling RIB.Upsert in the DataPlane.ForEach
callback which calls DataPlane.ProcessUpdate internally. That causes
deadlock in the SRv6 DataPlane. Fix it by moving RIB.Upsert out of
DataPlane.ForEach callback.

We couldn't notice this in the unit test because the fake dataplane we
are using in the test doesn't implement this RW locking. Implement it to
prevent the same issue in the future.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>