ci: add action to retrieve wolfi packages
This action will be based on the fact x86 and arm64 packages are released at the same time on wolfi repo, so we use the x86 packages as a base.
We pull the APKINDEX available at packages.wolfi.dev (official wolfi repo), untar, get the packages we want looking at the wolfi image yaml file,
and generate json package files in /tmp/wolfi-packages (directly in the runner) so renovate can exploit them using custom datasource.

Signed-off-by: Antony Reynaud <antony.reynaud@isovalent.com>